Move "Select parent" to "More options" menu.
