+++ packages/block-editor/src/components/block-toolbar/index.js
@@ -10,7 +10,6 @@ import { getBlockType, hasBlockSupport } from '@wordpress/blocks';
  * Internal dependencies
  */
 import BlockMover from '../block-mover';
-import BlockParentSelector from '../block-parent-selector';
 import BlockSwitcher from '../block-switcher';
 import BlockControls from '../block-controls';
 import BlockFormatControls from '../block-format-controls';
@@ -26,7 +25,6 @@ export default function BlockToolbar( {
 		blockClientIds,
 		blockClientId,
 		blockType,
-		hasParent,
 		hasReducedUI,
 		isValid,
 		isVisual,
@@ -34,7 +32,6 @@ export default function BlockToolbar( {
 		const {
 			getBlockName,
 			getBlockMode,
-			getBlockParents,
 			getSelectedBlockClientIds,
 			isBlockValid,
 			getBlockRootClientId,
@@ -53,7 +50,6 @@ export default function BlockToolbar( {
 				getBlockType( getBlockName( selectedBlockClientId ) ),
 			hasReducedUI: settings.hasReducedUI,
 			rootClientId: blockRootClientId,
-			hasParent: getBlockParents( selectedBlockClientId ).length > 0,
 			isValid: selectedBlockClientIds.every( ( id ) =>
 				isBlockValid( id )
 			),
@@ -95,11 +91,6 @@ export default function BlockToolbar( {
 
 	return (
 		<Wrapper className="block-editor-block-toolbar">
-			{ hasParent && ! isMultiToolbar && (
-				<ToolbarGroup className="block-editor-block-toolbar__block-parent-selector-wrapper">
-					<BlockParentSelector clientIds={ blockClientIds } />
-				</ToolbarGroup>
-			) }
 			<div ref={ nodeRef } { ...showBlockHighlightGestures }>
 				{ ( shouldShowVisualToolbar || isMultiToolbar ) && (
 					<ToolbarGroup className="block-editor-block-toolbar__block-controls">
