+++ packages/e2e-tests/specs/editor/various/toolbar-roving-tabindex.test.js
@@ -30,23 +30,6 @@ async function testBlockToolbarKeyboardNavigation( currentBlockLabel ) {
 	await expectLabelToHaveFocus( 'Move up' );
 }
 
-async function wrapCurrentBlockWithGroup() {
-	await page.click( '[aria-label="Change block type or style"]' );
-	await page.evaluate( () => {
-		document.querySelector( '.editor-block-list-item-group' ).click();
-	} );
-}
-
-async function testGroupKeyboardNavigation( currentBlockLabel ) {
-	await expectLabelToHaveFocus( 'Block: Group' );
-	await page.keyboard.press( 'Tab' );
-	await expectLabelToHaveFocus( currentBlockLabel );
-	await pressKeyWithModifier( 'shift', 'Tab' );
-	await expectLabelToHaveFocus( 'Select parent (Group)' );
-	await page.keyboard.press( 'ArrowRight' );
-	await expectLabelToHaveFocus( 'Change block type or style' );
-}
-
 describe( 'Toolbar roving tabindex', () => {
 	beforeEach( async () => {
 		await createNewPost();
@@ -58,31 +41,23 @@ describe( 'Toolbar roving tabindex', () => {
 		await insertBlock( 'Paragraph' );
 		await page.keyboard.type( 'Paragraph' );
 		await testBlockToolbarKeyboardNavigation( 'Paragraph block' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Paragraph block' );
 	} );
 
 	it( 'ensures heading block toolbar uses roving tabindex', async () => {
 		await insertBlock( 'Heading' );
 		await page.keyboard.type( 'Heading' );
 		await testBlockToolbarKeyboardNavigation( 'Block: Heading' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Block: Heading' );
 	} );
 
 	it( 'ensures list block toolbar uses roving tabindex', async () => {
 		await insertBlock( 'List' );
 		await page.keyboard.type( 'List' );
 		await testBlockToolbarKeyboardNavigation( 'Block: List' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Block: List' );
 	} );
 
 	it( 'ensures image block toolbar uses roving tabindex', async () => {
 		await insertBlock( 'Image' );
 		await testBlockToolbarKeyboardNavigation( 'Block: Image' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Block: Image' );
 	} );
 
 	it( 'ensures table block toolbar uses roving tabindex', async () => {
@@ -93,14 +68,10 @@ describe( 'Toolbar roving tabindex', () => {
 		await expectLabelToHaveFocus( 'Change block type or style' );
 		await page.click( '.blocks-table__placeholder-button' );
 		await testBlockToolbarKeyboardNavigation( 'Block: Table' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Block: Table' );
 	} );
 
 	it( 'ensures custom html block toolbar uses roving tabindex', async () => {
 		await insertBlock( 'Custom HTML' );
 		await testBlockToolbarKeyboardNavigation( 'Block: Custom HTML' );
-		await wrapCurrentBlockWithGroup();
-		await testGroupKeyboardNavigation( 'Block: Custom HTML' );
 	} );
 } );
